Replace the cache class with iodict
IODict is a library developed under the directord organization and
serves as a replacement for the current Cache class. While the Cache
class and IODict are similar, IODict implements the whole python dict
API, and provides 100% test coverage.

IODict is a new dependency, from within the Directord organization. This
will require additional packaging efforts, but will provide a cleaner
code base within directord and serve as a starting point to close the
durable queue efforts related to issue #343.
