+++ setup.py
@@ -51,6 +51,7 @@ setuptools.setup(
         "ssh-python",
         "tabulate",
         "tenacity",
+        "iodict",
     ],
     long_description=long_description,
     long_description_content_type="text/markdown",
